Refactor email into it's own middleware
